inprogress6-python-password-management/requirements.txt
anyio==4.6.2.post1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
graphql-core==3.2.5
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
promise==2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
Rx==1.6.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
strawberry-graphql==0.247.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4